moved load_lineage_and_acc2tax_maps to taxonomy.rs
